[Silver IV] Title: 피하자, Time: 148 ms, Memory: 14088 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/25379. 피하자/피하자.cc
@@ -30,8 +30,8 @@ int main(void)
     
     vector<int>v2 = v;
 
-    int cnt1 = 0;
-    int cnt2 = 0;
+    long long cnt1 = 0;
+    long long cnt2 = 0;
 
     queue<int> temp;
